+++ SOUP.user.js
@@ -3,7 +3,7 @@
 // @namespace   https://github.com/vyznev/
 // @description Miscellaneous client-side fixes for bugs on Stack Exchange sites (development)
 // @author      Ilmari Karonen
-// @version     1.41.3
+// @version     1.41.4
 // @copyright   2014-2015, Ilmari Karonen (http://stackapps.com/users/10283/ilmari-karonen)
 // @license     ISC; http://opensource.org/licenses/ISC
 // @match       *://*.stackexchange.com/*
@@ -260,7 +260,7 @@ fixes.math12803b = {
 fixes.codegolf959 = {
 	title:	"Add line-height shortener to the ascii-art tag",
 	url:	"http://meta.codegolf.stackexchange.com/q/959",
-	sites:	/^(meta\.)?codegolf\./,
+	sites:	/^(meta\.)?(codegolf|puzzling)\./,
 	css:	"pre { line-height: 1.15 }"
 };
 fixes.mse229797 = {