+++ CHANGELOG.md
@@ -20,6 +20,7 @@ SOUP version numbering generally follows the _major.minor.patch_ scheme.  Odd mi
 **Other changes:**
 
 * Don't even try to run late setup and normal script fixes unless jQuery is available; it won't work.  Early (and mathjax) script fixes still run.
+* [mse22372](http://meta.stackexchange.com/q/22372): Don't rewrite links to per-site metas (meta.*.stackexchange.com) to use HTTPS; they [currently return a 430 Forbidden error.](http://meta.stackexchange.com/q/265918)
 
 1.40 (25 Oct 2015)
 ====